+++ TCTC/TCTC/TCTC.cs
@@ -30,7 +30,7 @@ namespace TCTC
         public static TCTCards instance { get; private set; }
         private const string ModId = "TheCampingTurret.Rounds.TCTC.cards";
         private const string ModName = "TCTCards";
-        public const string Version = "2.0.0"; // What version are we on (major.minor.patch)?
+        public const string Version = "2.0.1"; // What version are we on (major.minor.patch)?
         public const string ModInitials = "TCTC";
 
         private static readonly AssetBundle Bundle = Jotunn.Utils.AssetUtils.LoadAssetBundleFromResources("tctc", typeof(TCTCards).Assembly);